bug fix follow/unfollow function
