+++ js/feed.js
@@ -1,6 +1,6 @@
-// import { apiFetch } from "./utils.js";
+import { apiFetch } from "./utils.js";
 import { buildFeed } from "./components/postHandler.mjs";
-// import { baseUrl, postUrl, authorInclude } from "./API/urls.js";
+import { baseUrl, postUrl, authorInclude } from "./API/urls.js";
 import { navListeners, signOutListener } from "./components/navLinks.mjs";
 import { feedFilter } from "./components/searchHandler.mjs";
 
@@ -8,4 +8,3 @@ navListeners();
 signOutListener();
 
 buildFeed();
-feedFilter();
